Restructure package and update entry point
