+++ live_caption/__init__.py
@@ -1,0 +1,5 @@
+"""Live Caption package."""
+
+from .cli import main, run_app
+
+__all__ = ["main", "run_app"]
