+++ live_caption/recognizer.py
